requirements.txt
asn1crypto==0.22.0
beautifulsoup4==4.6.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
idna==2.5
nltk==3.2.4
py==1.4.34
pycparser==2.18
pycrypto==2.6.1
pytest==3.1.3
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.12
tornado==4.5.1
urllib3==1.21.1
